Revert "refactor: move s2 test to a dedicated module"
This reverts commit 98e0b6f7b5bed8b0bffa5f0a0c743241a8f1e378.
