+++ tests/test_frbc_utils.py
@@ -3,15 +3,8 @@ from __future__ import annotations
 import math
 
 import pytest
-
-try:
-    from s2python.common import CommodityQuantity, NumberRange, PowerRange
-    from s2python.frbc import FRBCOperationMode, FRBCOperationModeElement
-except ImportError:
-    raise ImportError(
-        "The 's2-python' package is required for this functionality. "
-        "Install it using `pip install flexmeasures-client[s2]`."
-    )
+from s2python.common import CommodityQuantity, NumberRange, PowerRange
+from s2python.frbc import FRBCOperationMode, FRBCOperationModeElement
 
 from flexmeasures_client.s2.control_types.FRBC.utils import (
     compute_factor,
